teste 1/myapp/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
python-dotenv
requests
pymysql